+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2015-05-28
 * added put
 * replaced /etc/joowar-config.php with config.php in root installation
+* tested on server
 
 2015-05-17
 * first draft: list, get